▣ Inclusion of internal-specification logic.
● Key points:

• The internal-specification logic was builded for cases that you need to introduce a element that is not a pattern or can caught in current logic.

An example is the count, that need to execute this:

COUNT(*) OVER ()

For achive this and other case, I create the Standalone and Collective ways to introduce a non-pattern field in the response.

Standalone => Create the field outside of the data.
Collective => Create the field in each row of data.

Actual query example:

SELECT
[t1].[date_action] AS [f1], [t1].[id_action] AS [f2], [t1].[id_user] AS [f3], [t1].[id_action] AS [f4], [t1].[id_relation] AS [f5], [t1].[id_topic] AS [f6], COUNT(*) OVER () AS [is-s1]
FROM
[Action] AS [t1]
WHERE
((([t1].[id_action]) = 1) AND ([t1].[id_action]) = 2)

URL:

http://localhost:8900/api/action?$filter=((action_id eq 1) and action_id eq 2)&$compute=id_action as action_id&$count=true